+++ behavex/progress_bar.py
@@ -1,6 +1,8 @@
 import sys
 import time
 
+from behavex.global_vars import global_vars
+
 
 class ProgressBar:
     def __init__(self, prefix, total, bar_length=20, print_updates_in_new_lines=False):
@@ -8,7 +10,7 @@ class ProgressBar:
         self.total = total
         self.bar_length = bar_length
         self.current_iteration = 0
-        self.start_time = time.time()
+        self.start_time = global_vars.execution_start_time
         self.print_in_new_lines = print_updates_in_new_lines
 
     def start(self, start_increment=0):
@@ -33,7 +35,7 @@ class ProgressBar:
             percent = 100 * (self.current_iteration / float(self.total))
             filled_length = int(self.bar_length * self.current_iteration // self.total)
         bar = '█' * filled_length + '-' * (self.bar_length - filled_length)
-        elapsed_time = time.time() - self.start_time
+        elapsed_time = global_vars.execution_elapsed_time
         elapsed_formatted = time.strftime("%M:%S", time.gmtime(elapsed_time))
         progress_bar_content = f"\r{prefix}{percent:.0f}%|{bar}| {self.current_iteration}/{self.total} [{elapsed_formatted}]\r"
         if self.print_in_new_lines or new_line or percent == 100: